requirements.txt
torch
transformers==4.43.3
trl==0.9.6
openai
peft
datasets
pandas
jinja2
tqdm
pyyaml
gradio_client
python-dotenv
wheel
python-dotenv
tensorboard
numpy
alfworld[full]
sentencepiece
wandb
requests
google-generativeai
anthropic